+++ script/cleanup.sh
@@ -1,5 +1,4 @@
 #!/bin/bash -eux
-
 echo "==> Clear out machine id"
 rm -f /etc/machine-id
 touch /etc/machine-id
@@ -10,44 +9,29 @@ if grep -q -i "release 6" /etc/redhat-release ; then
     rm -f /etc/udev/rules.d/70-persistent-net.rules
     mkdir /etc/udev/rules.d/70-persistent-net.rules
     rm /lib/udev/rules.d/75-persistent-net-generator.rules
+
+    for ndev in `ls -1 /etc/sysconfig/network-scripts/ifcfg-*`; do
+    if [ "`basename $ndev`" != "ifcfg-lo" ]; then
+        sed -i '/^HWADDR/d' "$ndev";
+        sed -i '/^UUID/d' "$ndev";
+    fi
+    done
 fi
 rm -rf /dev/.udev/
-if [ -f /etc/sysconfig/network-scripts/ifcfg-eth0 ] ; then
-    sed -i "/^HWADDR/d" /etc/sysconfig/network-scripts/ifcfg-eth0
-    sed -i "/^UUID/d" /etc/sysconfig/network-scripts/ifcfg-eth0
-fi
-
-# Fix for https://github.com/CentOS/sig-cloud-instance-build/issues/38
-cat > /etc/sysconfig/network-scripts/ifcfg-eth0 << EOF
-DEVICE="eth0"
-BOOTPROTO="dhcp"
-ONBOOT="yes"
-TYPE="Ethernet"
-PERSISTENT_DHCLIENT="yes"
-EOF
+# if [ -f /etc/sysconfig/network-scripts/ifcfg-eth0 ] ; then
+#     sed -i "/^HWADDR/d" /etc/sysconfig/network-scripts/ifcfg-eth0
+#     sed -i "/^UUID/d" /etc/sysconfig/network-scripts/ifcfg-eth0
+# fi
 
 DISK_USAGE_BEFORE_CLEANUP=$(df -h)
 
-# Other locales will be removed from the VM
-KEEP_LANGUAGE="en"
-KEEP_LOCALE="en_US"
-echo "==> Remove unused man page locales"
-pushd /usr/share/man
-if [ $(ls | wc -w) -gt 16 ]; then
-  mkdir ../tmp_dir
-  mv man* $KEEP_LANGUAGE $SECONDARY_LANGUAGE ../tmp_dir
-  rm -rf *
-  mv ../tmp_dir/* .
-  rm -rf ../tmp_dir
-  sync
+if [[ $CLEANUP_BUILD_TOOLS  =~ true || $CLEANUP_BUILD_TOOLS =~ 1 || $CLEANUP_BUILD_TOOLS =~ yes ]]; then
+    echo "==> Removing tools used to build virtual machine drivers"
+    yum -y remove gcc libmpc mpfr cpp kernel-devel kernel-headers
 fi
-popd
 
 echo "==> Clean up yum cache of metadata and packages to save space"
 yum -y --enablerepo='*' clean all
-
-echo "==> Clear core files"
-rm -f /core*
 
 echo "==> Removing temporary files used to build box"
 rm -rf /tmp/*
@@ -55,6 +39,9 @@ rm -rf /tmp/*
 echo "==> Rebuild RPM DB"
 rpmdb --rebuilddb
 rm -f /var/lib/rpm/__db*
+
+# delete any logs that have built up during the install
+find /var/log/ -name *.log -exec rm -f {} \;
 
 echo '==> Clear out swap and disable until reboot'
 set +e
@@ -84,7 +71,7 @@ rm -f /EMPTY
 sync
 
 echo "==> Disk usage before cleanup"
-echo ${DISK_USAGE_BEFORE_CLEANUP}
+echo "${DISK_USAGE_BEFORE_CLEANUP}"
 
 echo "==> Disk usage after cleanup"
 df -h