+++ script/parallels.sh
@@ -10,7 +10,4 @@ if [[ $PACKER_BUILDER_TYPE =~ parallels ]]; then
     umount /mnt
     rm -rf $SSH_USER_HOME/prl-tools-lin.iso
     rm -f $SSH_USER_HOME/.prlctl_version
-
-    echo "==> Removing packages needed for building guest tools"
-    yum -y remove gcc cpp libmpc mpfr kernel-devel kernel-headers perl || echo "yum exit code $? is suppressed"
 fi